flex/requirements.txt
webapp2
Paste
WebOb
pandas
google-cloud-datastore
gunicorn